Fix NextWord in ModifiedKnesseNey
It was checking the first 2 strings, not the last 2.
Evaluate was missing out on some possible backup estimation. It cheats and returns values that aren't the real smoothed probability of the n-gram, if it doesn't contain some of the prefix strings. This works for our word prediction usage because the prefixes will be the same for all guesses.